requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
idna==3.7
overpy==0.7
psycopg==3.1.19
psycopg-binary==3.1.19
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
